--- src/js/components/Task.js
@@ -1,26 +1,0 @@
-import { TASK_DELAY, EXPLIAN_CODE } from "../utils/constans.js";
-
-class Task {
-	#$excutionBox;
-	#$explainBox;
-
-	constructor(className) {
-		this.#$excutionBox = document.querySelector(`.${className}`);
-		this.#$explainBox = document.querySelector(".explain");
-	}
-
-	removeTask() {
-		this.#$excutionBox.innerHTML = "";
-		this.#$explainBox.innerHTML = "";
-	}
-
-	appendTask(calleeName) {
-		this.#$excutionBox.innerHTML = `<div class="task center">${calleeName}</div>`;
-		this.#$explainBox.innerHTML = EXPLIAN_CODE[calleeName];
-		setTimeout(() => {
-			this.removeTask();
-		}, TASK_DELAY);
-	}
-}
-
-export default Task;
